% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{as.hourly}
\alias{as.hourly}
\title{An Index Class for Hourly Data}
\usage{
as.hourly(x, tz = ri_timezone())
}
\arguments{
\item{x}{object of class POSIXct}

\item{tz}{character, time zone. Defaults to \code{ri_timezone()}.}
}
\value{
POSIXct object of same length as \code{x} but with
ourly data (used for indexing in aggregation).
}
\description{
An Index Class for Hourly Data
}
