% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{calc_trip}
\alias{calc_trip}
\title{Calculate from-to geometry and time for the way the
bike has travelled between two points we've seen it.}
\usage{
calc_trip(x)
}
\description{
Calculate from-to geometry and time for the way the
bike has travelled between two points we've seen it.
}
\author{
Reto
}
