% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{ri_bbox}
\alias{ri_bbox}
\title{Stadtrad Innsbruck Bounding Box}
\usage{
ri_bbox(xmin = 11.3, xmax = 11.5, ymin = 47.2, ymax = 47.3)
}
\arguments{
\item{xmin, xmax, ymin, ymax}{single numeric, defining limits for longitude
(\code{x}) and latitude (\code{y}). Minimum must be smaller than the maximum.}
}
\value{
Object of class \code{bbox} with CRS $4326$.
}
\description{
Bounding box used to delete locations (places) outside range.
Defaults to $11.3-11.5$E, $47.2-47.3$N but can be changed if needed.
}
\author{
Reto
}
