% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes.R
\name{ri_bikes}
\alias{ri_bikes}
\alias{plot.ri_bikes}
\title{Bikes with Rudimentary Stats}
\usage{
ri_bikes(con)

\method{plot}{ri_bikes}(x, ...)
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{x}{object of class \code{ri_bikes}.}

\item{\dots}{forwarded to plot method.}
}
\value{
Spatial feature data.frame containing all ever recorded bikes
(identified by bike number), number of records in different places,
and the average position of the bike.
\itemize{
\item \code{number}: Bike identifier (bike number).
\item \code{records}: Number of stored records the bike has been seen in
different locations (i.e., moved places).
\item \code{avg_time_h}: Average parking time in hours.
\item \code{geometry}: Unweighted average longitude and latitude of all
the places the bike has been recorded.
}
}
\description{
Returns all bikes (by \code{number}) ever seen with some rudimentary stats.
}
\examples{
\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- ri_bikes(con)
head(x)
plot(x)
}

}
\author{
Reto
}
