% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes.R
\name{ri_biketrips}
\alias{ri_biketrips}
\title{Calculate Bike Trips}
\usage{
ri_biketrips(x)
}
\arguments{
\item{x}{object of class \code{ri_bikes}.}
}
\description{
The object returned by \code{ri_bikes()} contains the stops the bikes
have been recorded.
}
\author{
reto
}
