% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{ri_get_bike_numbers}
\alias{ri_get_bike_numbers}
\title{Parsing JSON Data}
\usage{
ri_get_bike_numbers(files, mc.cores = pmax(1L, detectCores() - 2L))
}
\arguments{
\item{files}{can be a ZIP file containing a series of JSON files, or
a character vector with one or multiple json files. If ZIP,
the ZIP file is unpacked and all JSON files found within are
processed.}

\item{mc.cores}{number of cores, defaults to all but two cores if not
specified otherwise.}
}
\value{
Returns an integer vector with all unique bike IDs seen
in the file/files.
}
\description{
Extracting all bike numbers seen.
}
