% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadtteile.R
\name{ri_innsbruck}
\alias{ri_innsbruck}
\title{Stadtteile Innsbruck}
\usage{
ri_innsbruck(...)
}
\arguments{
\item{...}{unused.}
}
\value{
Returns a single \code{sfc_POLYGON} used to subset bike positions
in \code{\link[=ri_trips]{ri_trips()}}.
}
\description{
Simple feature data frame with the districts of the city of Innsbruck.
}
