% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ri_parse_json}
\alias{ri_parse_json}
\title{Parsing JSON Files}
\usage{
ri_parse_json(x, domain = "si", verbose = TRUE)
}
\arguments{
\item{x}{character name of the file(s) to be parsed.}

\item{domain}{character, defaults to \code{"si"}. Ensures the input
files are for a specific domain (\code{"si"} = Stadtrad Innsbruck).}

\item{verbose}{logical. If \code{TRUE} some messages will be printed.}
}
\value{
Named list with two elements containing data.frames with
the \code{places} information and the \code{bikes} information in a raw format.
\code{bikes$datetime} is extracted from the JSON file name.
}
\description{
Helper function to parse the original JSON files, or ZIP archives
with these JSON files.
}
\details{
If \code{x} is a ZIP archive it will be unpacked and all JSON files
inside will be parsed. The file names must match the expected pattern
as we extract the time stamp from the file name.
}
\author{
Reto
}
