% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R, R/routes.R
\name{ri_places}
\alias{ri_places}
\alias{plot.ri_places}
\alias{ri_routes}
\title{Retrieve Places (Locations)}
\usage{
ri_places(con)

\method{plot}{ri_places}(x, bbox = ri_bbox(), ...)

ri_routes(con)
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{x}{object of class \code{ri_places}.}

\item{bbox}{object of class \code{bbox}.}

\item{\dots}{forwarded to plot method.}
}
\description{
The API reports where (how many) bikes are located when parked.
Besides around 50 official stations (as of 2025) bikes can also
be parked outside official stations.

Calculates routes (bikes moved from A to B).
}
\examples{
\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- ri_places(con)
head(x)
plot(x)
}

\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- ri_routes(con)
head(x)
plot(x)
}

}
\author{
Reto
}
