% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ri_read_jsons}
\alias{ri_read_jsons}
\alias{ri_read_json}
\alias{ri_read_zip}
\alias{ri_read_zips}
\alias{plot.ri_data}
\title{Reading Stadtrad Data}
\usage{
ri_read_jsons(
  x,
  city_name = "Innsbruck",
  cores = NULL,
  tz = NULL,
  verbose = FALSE
)

ri_read_json(
  x,
  city_name = "Innsbruck",
  returnclass = c("data.frame", "list"),
  tz = NULL,
  verbose = FALSE
)

ri_read_zip(
  x,
  city_name = "Innsbruck",
  cores = NULL,
  tz = NULL,
  verbose = FALSE
)

ri_read_zips(
  x,
  city_name = "Innsbruck",
  cores = NULL,
  tz = NULL,
  verbose = FALSE
)

\method{plot}{ri_data}(x, what = c("cities", "places"), ...)
}
\arguments{
\item{x}{character, name/path of the file(s) to be read. See Section
Details for more information.}

\item{city_name}{character of length one, defaults to \code{"Innsbruck"}. Used to
subset the data to one specific city.}

\item{cores}{\code{NULL} (default) or integer, not used by \code{ri_read_json}.
Number of cores to used with \code{parallel::mclapply}. If \code{NULL} it takes all but
two available cores (topped of by the maximum number of json files to be
processed).}

\item{tz}{defaults to \code{NULL}, forwarded to \code{\link[=ri_timezone]{ri_timezone()}}.}

\item{verbose}{integer, defaults to \code{0L} (\code{FALSE}; silent). If it evaluates to
\code{1L} (\code{TRUE}) it is verbose on function level, \code{2L} makes the interfaced
functions verbose, too.}

\item{returnclass}{only used by \code{ri_read_json}, one of "\verb{data.frame"}
(tibble data frames, default) or \code{"list"} (only for \code{ri_read_json}).}

\item{what}{what to plot.}

\item{...}{currently unused.}
}
\value{
Object of class \code{ri_data}, a named list with three tibble data
frames: \code{cities}, \code{places}, and \code{bikes}. The date and time information
extracted from the file name(s) \code{x} is added as a new variable \code{datetime}.
All data frames are sorted by datetime (univariate).

There is one exception: Calling \code{ri_read_json} allows to set \code{returnclass = "list"}
in which case the return is a named list of length four with mostly raw data,
with the elements \code{cities}, \code{places}, and \code{bikes} as for all the other functions,
except they do not include the datetime information. Instead, the datetime information
is stored on the fourth element \code{datetime}.
}
\description{
The REST API returns a single JSON file at the time of the request.
Individual JSON files can be processed using \code{ri_read_json}. Note that
the name of the file must include an integer, the timestamp of the request.
A series of convenience functions exist to read not only one, but multiple
files at a time. \code{ri_read_jsons} imports a series of JSON files and returns
a combined data set. If archived in a ZIP file, \code{ri_read_zip} will read all
available JSON files from that archive. Multiple archive files can be read
using the \code{ri_read_zips} function.
}
\examples{
\dontrun{
## Reading JSON files:
## Find available json files
jsonfiles <- list.jsonfiles(file.path("archive", "foo"), full.names = TRUE)
jsonfiles <- jsonfiles[grepl("^[0-9]+_si.json$", basename(jsonfiles))]

## Reading single json file
x <- ri_read_json(jsonfiles[1], verbose = TRUE)

## Reading set of json files
y <- ri_read_jsons(jsonfiles[1:3], verbose = TRUE)

## Quick visual inspection
library("tinyplot")
tinyplot(available_bikes ~ datetime, data = y$cities)
tinyplot(bikes ~ datetime | name, type = "l",
         data = subset(y$places, !grepl("^BIKE", name)))

## Reading ZIP archive files
## Find available ZIP files
zipfiles <- list.files("archive", full.names = TRUE)
zipfiles <- zipfiles[grepl("_si\\\\.zip$", zipfiles)]

#' ## Reading single ZIP file
x <- ri_read_zip(zipfiles[1], verbose = TRUE)

#' ## Reading set of ZIP files
y <- ri_read_zips(zipfiles[1:5], verbose = TRUE)

## Quick visual inspection
tinyplot(available_bikes ~ datetime | as.Date(datetime),
         data = y$cities, type = "l")
}

}
\author{
Reto
}
