% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentals.R
\name{ri_rentals}
\alias{ri_rentals}
\alias{plot.ri_rentals}
\title{Availability of Rentals}
\usage{
ri_rentals(con, officials_only = TRUE)

\method{plot}{ri_rentals}(x, type = 2, bbox = ri_bbox(), ...)
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{officials_only}{logical, defaults to \code{TRUE}. If \code{FALSE}, random
parking locations are also included.}

\item{x}{object of class \code{ri_rentals}.}

\item{type}{integer, type of plot (currently \code{1}, \code{2}).}

\item{bbox}{object of class \code{bbox}.}

\item{\dots}{forwarded to plot method.}
}
\description{
Every few minutes the number of available bikes per station
is recorded. This is mainly useful for official stations, thus
`random locations' are excluded by default.
}
\examples{
\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- ri_rentals(con)
head(x)
plot(x)
}

}
\author{
Reto
}
