% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{ri_timezone}
\alias{ri_timezone}
\title{Get and Set Default Time Zone}
\usage{
ri_timezone(x = NULL)
}
\arguments{
\item{x}{\code{NULL} or character (see \code{?OlsonNames}). If set, the
default (\code{"Europe/London"}) will be overwritten and used for the
current session.}
}
\value{
Character with current time zone used by the package.
}
\description{
The package will handle times in local time zone, defaulting to
'Europe/Vienna'. This function allows to set the time zone
and/or retrieve the time zone.
}
