% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes.R
\name{si_bikes}
\alias{si_bikes}
\alias{plot.si_bikes}
\title{Bikes with Rudimentary Stats}
\usage{
si_bikes(con)

\method{plot}{si_bikes}(x, ...)
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{x}{object of class \code{si_bikes}.}

\item{\dots}{forwarded to plot method.}
}
\description{
Returns all bikes (by \code{number}) ever seen with some rudimentary stats.
}
\examples{
\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- si_bikes(con)
head(x)
plot(x)
}

}
\author{
Reto
}
