% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{si_places}
\alias{si_places}
\alias{plot.si_places}
\title{Retrieve Places (Locations)}
\usage{
si_places(con)

\method{plot}{si_places}(x, bbox = si_bbox(), ...)
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{x}{object of class \code{si_places}.}

\item{bbox}{object of class \code{bbox}.}

\item{\dots}{forwarded to plot method.}
}
\description{
The API reports where (how many) bikes are located when parked.
Besides around 50 official stations (as of 2025) bikes can also
be parked outside official stations.
}
\examples{
\dontrun{
library('RSQLite')
con <- dbConnect(SQLite(), "../stadtrad.db")
x <- si_places(con)
head(x)
plot(x)
}

}
\author{
Reto
}
