% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stadtteile}
\alias{stadtteile}
\title{Stadtteile}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 3 columns.
}
\usage{
stadtteile
}
\description{
A simple feature data.frame containing the (main) districts
of Innsbruck.
}
\examples{
data(stadtteile)
summary(stadtteile)

}
\author{
Reto
}
\keyword{datasets}
