% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes.R
\name{ri_bikes}
\alias{ri_bikes}
\title{Loading Bike Data}
\usage{
ri_bikes(con, tz = ri_timezone())
}
\arguments{
\item{con}{database connection. Typically an \code{SQLiteConnection}
but should be possible to easily extend to MariaDB/MySQL (untested).}

\item{tz}{chracter, time zone. Defaults to \code{ri_timezone()}.}
}
\value{
Data frame with details about bikes locations (see 'Details').
}
\description{
Loading Bike Data
}
\details{
The object returned contains information about all recorded bike stops, i.e.,
when we have seen a certain bike at a specific place.
\itemize{
\item \code{first_seen}: Date and time he bike has first seen at the station.
\item \code{last_seen}: Date and time the bike has last been seen at the station.
\item \code{number}: Bike identifier.
\item \code{bike_type}: Type of bike.
\item \code{active}: Status.
\item \code{name}: Name of the station the bike is located.
\item \code{lon}/\code{lat}: Location of the bike (station/place).
}
}
\author{
reto
}
