% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ri_read_jsons}
\alias{ri_read_jsons}
\alias{ri_read_json}
\alias{ri_read_zip}
\alias{ri_read_zips}
\title{Reading Stadtrad Data}
\usage{
ri_read_jsons(
  x,
  city_name = "Innsbruck",
  cores = NULL,
  tz = NULL,
  verbose = FALSE
)

ri_read_json(
  x,
  city_name = "Innsbruck",
  returnclass = c("data.frame", "list"),
  tz = NULL,
  verbose = FALSE
)

ri_read_zip(x, city_name = "Innsbruck", tz = NULL, verbose = FALSE)

ri_read_zips(x, city_name = "Innsbruck", tz = NULL, verbose = FALSE)
}
\arguments{
\item{x}{character, name/path of the file(s) to be read. See Section
Details for more information.}

\item{city_name}{character of length one, defaults to \code{"Innsbruck"}. Used to
subset the data to one specific city.}

\item{cores}{\code{NULL} (default) or integer, not used by \code{ri_read_json}.
Number of cores to used with \code{parallel::mclapply}. If \code{NULL} it takes all but
two available cores (topped of by the maximum number of json files to be
processed).}

\item{tz}{defaults to \code{NULL}, forwarded to \code{\link[=ri_timezone]{ri_timezone()}}.}

\item{verbose}{integer, defaults to \code{0L} (\code{FALSE}; silent). If it evaluates to
\code{1L} (\code{TRUE}) it is verbose on function level, \code{2L} makes the interfaced
functions verbose, too.}

\item{returnclass}{only used by \code{ri_read_json}, one of "\verb{data.frame"}
(tibble data frames, default) or \code{"list"} (only for \code{ri_read_json}).}
}
\value{
A named list is returned which by default contains three
tibble data frames, namely \code{cities}, \code{places}, and \code{bikes}. The
date and time information extracted from the file name(s) \code{x} is
added as a new variable \code{datetime}.

Calling \code{ri_read_json} with \code{returnclass = "list"} the return is also
a named list but with four elements containing the raw data from the json
file subsetted to the city requested, plus one additional element
\code{datetime} with the timestamp (POSIXt).
}
\description{
The REST API returns a single JSON file at the time of the request.
Individual JSON files can be processed using \code{ri_read_json}. Note that
the name of the file must include an integer, the timestamp of the request.
A series of convenience functions exist to read not only one, but multiple
files at a time. \code{ri_read_jsons} imports a series of JSON files and returns
a combined data set. If archived in a ZIP file, \code{ri_read_zip} will read all
available JSON files from that archive. Multiple archive files can be read
using the \code{ri_read_zips} function.
}
\examples{
\dontrun{
## Find available json files
jsonfiles <- list.jsonfiles(file.path("archive", "foo"), full.names = TRUE)
jsonfiles <- jsonfiles[grepl("^[0-9]+_si.json$", basename(jsonfiles))]

## Reading single json file
x <- ri_read_json(jsonfiles[1], verbose = TRUE)

## Reading set of json files
y <- ri_read_jsons(jsonfiles[1:3], verbose = TRUE)

## Quick check
library("tinyplot")
tinyplot(available_bikes ~ datetime, data = x$cities)
tinyplot(bikes ~ datetime | name, type = "l",
         data = subset(x$places, !grepl("^BIKE", name)))
}

}
\author{
Reto
}
