% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{ri_trips}
\alias{ri_trips}
\title{Calculating Bike Trips}
\usage{
ri_trips(x, cores = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{ri_data} as returned by
\code{\link[=ri_read_jsons]{ri_read_jsons()}}, \code{\link[=ri_read_zip]{ri_read_zip()}}, or \code{\link[=ri_read_zips]{ri_read_zips()}}.}

\item{cores}{\code{NULL} (default) or integer, not used by \code{ri_read_json}.
Number of cores to used with \code{parallel::mclapply}. If \code{NULL} it takes all but
two available cores (topped of by the maximum number of json files to be
processed).}

\item{verbose}{defaults to \code{FALSE}.}

\item{...}{currently unused.}
}
\description{
Uses the raw data to calculate bike trips, i.e., calculate
when a bike disapeared at a certain location and then popped up
at another one (considered a rented trip).
}
\author{
Reto
}
