% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{ri_trips}
\alias{ri_trips}
\title{Calculating Bike Trips}
\usage{
ri_trips(x, cores = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{ri_data} as returned by
\code{\link[=ri_read_jsons]{ri_read_jsons()}}, \code{\link[=ri_read_zip]{ri_read_zip()}}, or \code{\link[=ri_read_zips]{ri_read_zips()}}.}

\item{cores}{\code{NULL} (default) or integer, not used by \code{ri_read_json}.
Number of cores to used with \code{parallel::mclapply}. If \code{NULL} it takes all but
two available cores (topped of by the maximum number of json files to be
processed).}

\item{verbose}{defaults to \code{FALSE}.}

\item{...}{currently unused.}
}
\value{
Object of class \code{c("ri_trips", "tbl", "data.frame")}
containing bike trips as well as 'open trips' (see below).
Besides the bike \code{number} (the identifier) the data frame
contains \verb{*_place_id}, \verb{*_daetetime}, \verb{*_name}, \verb{*_lng}, and \verb{*_lat}
for both the start and the end of the trip. In addition, the trip
duration is reported in minutes (rounded to closest minutes).

Besides closed trips (where we know the start and end) the data frame
also contains "open trips". These are always added at the start of
the data period (with respect to 'datetime') and end of the data period
and are required to capture potential bike trips which started before
the earliest observation into the period the data covers, as well
as trips starting within the data period but have not finished at
the end of the period.
These are later used when combining multiple \code{ri_trips} objects
(see S3 method \code{c.ri_trips}).
}
\description{
Uses the raw data to calculate bike trips, i.e., calculate
when a bike disapeared at a certain location and then popped up
at another one (considered a rented trip).
}
\author{
Reto
}
